packages/lol_vod_analyzer/pyproject.toml
[project]
name = "lol-vod-analyzer"
version = "0.1.0"
description = "LoL VOD analysis tool - extracts insights from coaching and gameplay videos"
requires-python = ">=3.12"
dependencies = [
    "yt-dlp>=2026.3",
    "Pillow>=10.0",
    "google-genai>=1.0",
    "typer[all]>=0.9",
    "rich>=13.7",
    "pydantic>=2.5",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "opencv-python>=4.8",
]

[project.scripts]
lol-vod = "lol_vod_analyzer.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/lol_vod_analyzer"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
]
